add 完成用户管理页面,添加拦截器,跨域请求配置 多环境上线

+++ src/main/java/com/qimu/smbms/model/request/BillPageRequest.java
@@ -15,7 +15,7 @@ public class BillPageRequest implements Serializable {
     Integer pageIndex;
     String productName;
     Integer isPayment;
-    Integer providerId;
+    Long providerId;
     Integer pageSize;
 
     public Integer getPageIndex() {
